Application/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
PyMySQL==1.1.0
cryptography==41.0.7
